git-svn-id: svn://svn.ijw.co.nz/svn/Profiler@1022 993157c7-ee19-0410-b2c4-bb4e9862e678

+++ ProfilerUi/FunctionNameProvider.cs
@@ -21,7 +21,7 @@ namespace ProfilerUi
 				if (m == null || !m.Success)
 					continue;
 
-				uint functionId = (uint)int.Parse(m.Groups[1].Value.Substring(2), NumberStyles.HexNumber);
+				uint functionId = uint.Parse(m.Groups[1].Value.Substring(2), NumberStyles.HexNumber);
 				string functionName = m.Groups[2].Value;
 
 				names.Add(functionId, functionName);
